# docs/context/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AFK Miners API (inferred)",
    "version": "0.0.1"
  },
  "paths": {
...
    "/list": {
      "get": {
        "responses": {
          "200": {
            "description": "OK (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error 400 (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Error 500 (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "try {\n      const playerId = req.user.id;\n      const { heroKey": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/select": {
      "post": {
        "responses": {
          "200": {
            "description": "OK (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error 400 (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Error 500 (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "try {\n      const playerId = req.user.id;\n      const { heroKey": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/status": {
      "get": {
        "responses": {
          "200": {
            "description": "OK (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Error 400 (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Error 500 (inferred)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "try {\n      const playerId = req.user.id;\n      const { heroKey": {
                    "type": "number"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}